pyproject.toml
[tool.poetry]
name = "hello"
version = "0.1.0"
description = ""
authors = ["pablo <pablohenriquem@hotmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0.6"
djangorestframework = "^3.15.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
